mcp-servers/zotero-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zotero-server"
version = "0.1.0"
description = "MCP server for Zotero Web API v3 reference management"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
zotero-server = "zotero_server:serve"

[tool.hatch.build.targets.wheel]
packages = ["src/zotero_server"]
